v11/requirements.txt
# V11 — Darvas Box + Volume Imbalance + LLM Filter
# Python 3.14+

# IBKR connection
ib_insync>=0.9.86
nest_asyncio>=1.6.0

# LLM client (OpenAI-compatible for Grok)
openai>=1.30.0

# Data validation
pydantic>=2.7.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
